Release/0.7.0

+++ scripts/app.sh
@@ -2,49 +2,32 @@
 # Required Notice: Copyright
 # Umbrel (https://umbrel.com)
 
+echo "Starting app script"
+
+source "${BASH_SOURCE%/*}/common.sh"
+
 set -euo pipefail
 
-# use greadlink instead of readlink on osx
-if [[ "$(uname)" == "Darwin" ]]; then
-  rdlk=greadlink
-else
-  rdlk=readlink
+ensure_pwd
+
+ROOT_FOLDER="${PWD}"
+STATE_FOLDER="${ROOT_FOLDER}/state"
+ENV_FILE="${ROOT_FOLDER}/.env"
+
+# Root folder in host system
+ROOT_FOLDER_HOST=$(grep -v '^#' "${ENV_FILE}" | xargs -n 1 | grep ROOT_FOLDER_HOST | cut -d '=' -f2)
+REPO_ID=$(grep -v '^#' "${ENV_FILE}" | xargs -n 1 | grep APPS_REPO_ID | cut -d '=' -f2)
+STORAGE_PATH=$(grep -v '^#' "${ENV_FILE}" | xargs -n 1 | grep STORAGE_PATH | cut -d '=' -f2)
+
+# Override vars with values from settings.json
+if [[ -f "${STATE_FOLDER}/settings.json" ]]; then
+  # If storagePath is set in settings.json, use it
+  if [[ "$(get_json_field "${STATE_FOLDER}/settings.json" storagePath)" != "null" ]]; then
+    STORAGE_PATH="$(get_json_field "${STATE_FOLDER}/settings.json" storagePath)"
+  fi
 fi
 
-ROOT_FOLDER="$($rdlk -f $(dirname "${BASH_SOURCE[0]}")/..)"
-REPO_ID="$(echo -n "https://github.com/meienberger/runtipi-appstore" | sha256sum | awk '{print $1}')"
-STATE_FOLDER="${ROOT_FOLDER}/state"
-
-show_help() {
-  cat <<EOF
-app 0.0.1
-
-CLI for managing Tipi apps
-
-Usage: app <command> <app> [<arguments>]
-
-Commands:
-    install                    Pulls down images for an app and starts it
-    uninstall                  Removes images and destroys all data for an app
-    stop                       Stops an installed app
-    start                      Starts an installed app
-    compose                    Passes all arguments to docker-compose
-    ls-installed               Lists installed apps
-EOF
-}
-
-# Get field from json file
-function get_json_field() {
-  local json_file="$1"
-  local field="$2"
-
-  echo $(jq -r ".${field}" "${json_file}")
-}
-
-list_installed_apps() {
-  str=$(get_json_field ${STATE_FOLDER}/apps.json installed)
-  echo $str
-}
+write_log "Running app script: ROOT_FOLDER=${ROOT_FOLDER}, ROOT_FOLDER_HOST=${ROOT_FOLDER_HOST}, REPO_ID=${REPO_ID}, STORAGE_PATH=${STORAGE_PATH}"
 
 if [ -z ${1+x} ]; then
   command=""
@@ -52,31 +35,10 @@ else
   command="$1"
 fi
 
-# Lists installed apps
-if [[ "$command" = "ls-installed" ]]; then
-  list_installed_apps
-
-  exit
-fi
-
 if [ -z ${2+x} ]; then
-  show_help
   exit 1
 else
-
   app="$2"
-  root_folder_host="${3:-$ROOT_FOLDER}"
-  repo_id="${4:-$REPO_ID}"
-
-  if [[ -z "${repo_id}" ]]; then
-    echo "Error: Repo id not provided"
-    exit 1
-  fi
-
-  if [[ -z "${root_folder_host}" ]]; then
-    echo "Error: Root folder not provided"
-    exit 1
-  fi
 
   app_dir="${ROOT_FOLDER}/apps/${app}"
 
@@ -84,36 +46,35 @@ else
     # copy from repo
     echo "Copying app from repo"
     mkdir -p "${app_dir}"
-    cp -r "${ROOT_FOLDER}/repos/${repo_id}/apps/${app}"/* "${app_dir}"
+    cp -r "${ROOT_FOLDER}/repos/${REPO_ID}/apps/${app}"/* "${app_dir}"
   fi
 
-  app_data_dir="${ROOT_FOLDER}/app-data/${app}"
+  app_data_dir="${STORAGE_PATH}/app-data/${app}"
 
   if [[ -z "${app}" ]] || [[ ! -d "${app_dir}" ]]; then
     echo "Error: \"${app}\" is not a valid app"
     exit 1
   fi
-
 fi
 
 if [ -z ${3+x} ]; then
   args=""
 else
-  args="${@:3}"
+  args="${*:3}"
 fi
 
 compose() {
   local app="${1}"
   shift
 
-  local architecture="$(uname -m)"
+  arch=$(uname -m)
+  local architecture="${arch}"
 
   if [[ "$architecture" == "aarch64" ]]; then
     architecture="arm64"
   fi
 
   # App data folder
-  local env_file="${ROOT_FOLDER}/.env"
   local app_compose_file="${app_dir}/docker-compose.yml"
 
   # Pick arm architecture if running on arm and if the app has a docker-compose.arm.yml file
@@ -121,19 +82,23 @@ compose() {
     app_compose_file="${app_dir}/docker-compose.arm.yml"
   fi
 
-  local common_compose_file="${ROOT_FOLDER}/repos/${repo_id}/apps/docker-compose.common.yml"
+  # Pick arm architecture if running on arm and if the app has a docker-compose.arm64.yml file
+  if [[ "$architecture" == "arm64" ]] && [[ -f "${app_dir}/docker-compose.arm64.yml" ]]; then
+    app_compose_file="${app_dir}/docker-compose.arm64.yml"
+  fi
+
+  local common_compose_file="${ROOT_FOLDER}/repos/${REPO_ID}/apps/docker-compose.common.yml"
 
   # Vars to use in compose file
-  export APP_DATA_DIR="${root_folder_host}/app-data/${app}"
-  export APP_DIR="${app_dir}"
-  export ROOT_FOLDER_HOST="${root_folder_host}"
-  export ROOT_FOLDER="${ROOT_FOLDER}"
+  export APP_DATA_DIR="${STORAGE_PATH}/app-data/${app}"
+  export ROOT_FOLDER_HOST="${ROOT_FOLDER_HOST}"
 
-  # Docker-compose does not support multiple env files
-  # --env-file "${env_file}" \
+  write_log "Running docker compose -f ${app_compose_file} -f ${common_compose_file} ${*}"
+  write_log "APP_DATA_DIR=${APP_DATA_DIR}"
+  write_log "ROOT_FOLDER_HOST=${ROOT_FOLDER_HOST}"
 
-  docker-compose \
-    --env-file "${ROOT_FOLDER}/app-data/${app}/app.env" \
+  docker compose \
+    --env-file "${app_data_dir}/app.env" \
     --project-name "${app}" \
     --file "${app_compose_file}" \
     --file "${common_compose_file}" \
@@ -142,7 +107,13 @@ compose() {
 
 # Install new app
 if [[ "$command" = "install" ]]; then
-  compose "${app}" pull
+  # Write to file script.log
+  write_log "Installing app ${app}..."
+
+  if ! compose "${app}" pull; then
+    write_log "Failed to pull app ${app}"
+    exit 1
+  fi
 
   # Copy default data dir to app data dir if it exists
   if [[ -d "${app_dir}/data" ]]; then
@@ -152,20 +123,30 @@ if [[ "$command" = "install" ]]; then
   # Remove all .gitkeep files from app data dir
   find "${app_data_dir}" -name ".gitkeep" -exec rm -f {} \;
 
-  chown -R "1000:1000" "${app_data_dir}"
+  chmod -R a+rwx "${app_data_dir}"
 
-  compose "${app}" up -d
-  exit
+  if ! compose "${app}" up -d; then
+    write_log "Failed to start app ${app}"
+    exit 1
+  fi
+
+  exit 0
 fi
 
 # Removes images and destroys all data for an app
 if [[ "$command" = "uninstall" ]]; then
-  echo "Removing images for app ${app}..."
+  write_log "Removing images for app ${app}..."
 
-  compose "${app}" up --detach
-  compose "${app}" down --rmi all --remove-orphans
+  if ! compose "${app}" up --detach; then
+    write_log "Failed to uninstall app ${app}"
+    exit 1
+  fi
+  if ! compose "${app}" down --rmi all --remove-orphans; then
+    write_log "Failed to uninstall app ${app}"
+    exit 1
+  fi
 
-  echo "Deleting app data for app ${app}..."
+  write_log "Deleting app data for app ${app}..."
   if [[ -d "${app_data_dir}" ]]; then
     rm -rf "${app_data_dir}"
   fi
@@ -174,14 +155,21 @@ if [[ "$command" = "uninstall" ]]; then
     rm -rf "${app_dir}"
   fi
 
-  echo "Successfully uninstalled app ${app}"
+  write_log "Successfully uninstalled app ${app}"
   exit
 fi
 
 # Update an app
 if [[ "$command" = "update" ]]; then
-  compose "${app}" up --detach
-  compose "${app}" down --rmi all --remove-orphans
+  if ! compose "${app}" up --detach; then
+    write_log "Failed to update app ${app}"
+    exit 1
+  fi
+
+  if ! compose "${app}" down --rmi all --remove-orphans; then
+    write_log "Failed to update app ${app}"
+    exit 1
+  fi
 
   # Remove app
   if [[ -d "${app_dir}" ]]; then
@@ -189,33 +177,41 @@ if [[ "$command" = "update" ]]; then
   fi
 
   # Copy app from repo
-  cp -r "${ROOT_FOLDER}/repos/${repo_id}/apps/${app}" "${app_dir}"
+  cp -r "${ROOT_FOLDER}/repos/${REPO_ID}/apps/${app}" "${app_dir}"
 
   compose "${app}" pull
-  exit
+  exit 0
 fi
 
 # Stops an installed app
 if [[ "$command" = "stop" ]]; then
-  echo "Stopping app ${app}..."
-  compose "${app}" rm --force --stop
-  exit
+  write_log "Stopping app ${app}..."
+
+  if ! compose "${app}" rm --force --stop; then
+    write_log "Failed to stop app ${app}"
+    exit 1
+  fi
+
+  exit 0
 fi
 
 # Starts an installed app
 if [[ "$command" = "start" ]]; then
-  echo "Starting app ${app}..."
-  compose "${app}" up --detach
-  exit
+  write_log "Starting app ${app}..."
+  if ! compose "${app}" up --detach; then
+    write_log "Failed to start app ${app}"
+    exit 1
+  fi
+  exit 0
 fi
 
-# Passes all arguments to docker-compose
+# Passes all arguments to Docker Compose
 if [[ "$command" = "compose" ]]; then
-  compose "${app}" ${args}
-  exit
+  if ! compose "${app}" "${args}"; then
+    write_log "Failed to run compose command for app ${app}"
+    exit 1
+  fi
+  exit 0
 fi
 
-# If we get here it means no valid command was supplied
-# Show help and exit
-show_help
 exit 1
